ticket: 5584
All exports in nidmgr32.dll are declared as __declspec(dllexport)
using the KHMEXP macro in the header files.  However, since the same
header files are used to declare imports when building applications
and plug-ins that use nidmgr32.dll, the KHMEXP macro should switch to
__declspec(import) so that the relevant import table entries are
created.

To make this switch, the source files that go into nidmgr32.dll are
compiled with the special macro _NIMLIB_ defined that indicates that
the KHMEXP should expand to __declspec(dllexport).  In the absence of
this macro, KHMEXP will expand to __declspec(dllimport).


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19618 dc483132-0cff-0310-8789-dd5450dbe970