ticket: 5584
Each error context in NIM can maintain a progress counter to indicate
the progress of the operation.  The progress counter for a context
that contains child contexts should be computed by adding the progress
counters of its children.  This update corrects the behavior of
kherr_get_progress() and kherr_get_progress_i() to take child contexts
into account.

In addition, since the progress counter of a context now depend on the
progress counters of its children, kherr_set_progress() will send
progress change notifications for any parent context as well.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19621 dc483132-0cff-0310-8789-dd5450dbe970