ticket: new subject: NIM: New command-line options --hide and --show / new command-line help dialog component: windows
Add two command-line options to the Network Identity Manager application:

--hide : If there is already an instance of Network Identity Manager
         running in the user's session, this will cause the main
         window of that instance to be hidden.

--show : Unhides the main window of the running instance.

If no options are specified when starting netidmgr.exe while another
instance is running, the running instance will perform the default
action as configured in the user's preferences.  By default, this will
be to show the main window.

The message box that displayed the command-line options if an invalid
option is specified has been replaced with a dialog box that shows the
options in a more readable rich text control.  The contents of the 
control are specified in ui/lang/<lang>/cmdline.rtf.  

Do not edit cmdline.rtf with Microsoft Word.  The output of Word is 
incompatible with the RichEdit20 component.




git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19854 dc483132-0cff-0310-8789-dd5450dbe970