ticket: new subject: NIM: Color Schemas component: windows
The default color scheme used by the Network Identity Manager credentials
display can sometimes cause the text to be difficult to read.  In
addition, since some of the colors are derived from colors used by the
current Windows theme, the selection, window background and text
colors may be mismatched with the colors that are hardcoded into the
application.

To rememdy this problem, this patch defines a new set of colors
that will be used with the credentials display.  The new scheme
doesn't use a special color to distinguish the default identity which
is already indicated with the text "(Default)" displayed alongside it.
Instead the colors are used to denote the remaining lifetime of
credentials and identities.

Also, the color scheme defines all the colors that it uses instead of
deriving some of them from the Windows color scheme.

All the color information is now kept in the Network Identity Manager
UI schema.  The schema automatically maps to the registry, so users
(and deployers) can override the colors by creating the necessary 
registry keys and values.  The registry keys containing color 
information are:

  Software\MIT\NetIDMgr\CredWindow\Themes and
  Software\MIT\NetIDMgr\CredWindow\Themes\Default

The "Default" key contains the default color scheme.  The color value
names are defined in ui\uiconfig.csv under Themes\_Schema
configuration subspace.  Each color is represented by a 32-bit number.
The low-order 24 bits contain a COLORREF value.  The high-order 8 bits
contain an alpha value which, if non-zero, will be used to blend the
color with the selection color (color value named ClrSelection).


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19855 dc483132-0cff-0310-8789-dd5450dbe970