ticket: new subject: NIM: Context menu selection issues component: windows
The handler for WM_CONTEXTMENU in the credentials view of Network
Identity Manager assumed that the context menu was invoked using the
mouse.  If it was, then the message parameters would specify the x and
y co-ordinates of the mouse.  The context menu can also be invoked via
the keyboard, in which case the x- and y- coordinates are set to
(-1,-1).

An additional problem with the code is that it was not selecting the
row that was right-clicked on if it was not already selected.  This
results in the some of the commands on the context menu targetting
credentials that the user didn't intend to target.

This patch adds support for handling the context menu when it is
invoked via the keyboard and also sets the selection to the row that
was right-clicked if the user invokes the context menu using the
mouse.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19856 dc483132-0cff-0310-8789-dd5450dbe970