ticket: new subject: NIM: do not permit resizing of filler columns component: windows
The credentials display of Network Identity Manager allows the user to
change the order and size of the displayed columns.  However, some
columns are not resizable.

An oversight in the earlier code allowed the user to change the size
of "filler" columns.  These columns are sized to fill up the remainder
of the display area after the other columns have been allocated.
Their width is determined by the size of the credentials display and
the width of the other columns.  They are used in the basic view where
the only columns are the flags column (fixed width) and the identity
column (filler).

This patch prevents the "filler" columns (columns with the
KHUI_CW_COL_FILLER flag) from being resized.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19858 dc483132-0cff-0310-8789-dd5450dbe970