ticket: new subject: NIM: remove unused code from ui/credwnd.c component: windows
Remove unused code from Network Identity Manager credwnd.c.  The code
was meant to construct a user interface context based on where the
user right clicks on the credentials display.  However, doing so
without indicating the changed selection to the user results in the
application performing an operation on an identity or credential that
the user didn't intend to select.

The code was commented out and was never used in any recent release of
Kerberos for Windows.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19861 dc483132-0cff-0310-8789-dd5450dbe970