ticket: new subject: NIM: support include files in schemas component: windows
The ccsv.pl and csvschema.cfg scripts are used to generate "C" source
code from CSV files containing tabular data.  In particular, these are 
used to define the configuration schema for Network Identity Manager 
and some of its plug-ins.

It is desirable to be able to include arbitrary header files and
define macros in the generated C code so that the schema definition
can use them.  This patch allows the CSV files to contain headers that
define lines of text that will be included literally in the generated
C code.  Lines at the start of schema CSV file that begin with '#@'
will be stripped of the '#@' prefix and inserted into the C code.

E.g: The following line at the start of a schema CSV file:

#@#include<windows.h>

,will result in the following text in the C code:

#include<windows.h>

Then the schema definition can use macros of the form:

ClrHeaderExpSel, KC_INT32, "RGB(195, 94, 94)"

,which use macros such as RGB that are defined in the included header
file.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19862 dc483132-0cff-0310-8789-dd5450dbe970