ticket: new subject: NIM: Reposition New Credentials Dialog if necessary component: windows
The new credentials dialog in Network Identity Manager does not check
whether it is positioned outside of the display screen.  It tries to
position itself in the center of the primary display if the Network
Identity Manager window is hidden and it tries to center itself over
the main window if the main window is visible.  If the main window is
too close to the edge of the screen, this may result in the new
credentials window being partially outside the display area.  This is
especially a problem when the new credentials dialog switches to the
advanced view.

The patch checks whether the window rectangle is visible before
repositioning the new credentials dialog and adjusts the window
rectangle so that it is.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19867 dc483132-0cff-0310-8789-dd5450dbe970