ticket: new subject: NIM: 64-bit Windows Support and Removal of Compile Time Warnings component: windows
This patch permits Network Identity Manager to be built for 64-bit Windows.
In the process all compile time warnings have been taken care of.  

For 64-bit Windows, we do not build the Kerberos v4 Credential Provider
and we will not attempt to load the krb524 library.  

Note that when testing the 64-bit NIM, there is no CCAPI at the 
moment so you must manually specify a FILE: ccache as part of the 
identity's Kerberos v5 configuration if you want to use cache's 
other than the MSLSA.

This patch also consolidates the computation of the default ccache 
name into utility functions:

  khm_krb5_get_identity_default_ccache
  khm_krb5_get_identity_default_ccacheA



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19891 dc483132-0cff-0310-8789-dd5450dbe970