ticket:new subject: krb5_fcc_generate_new is non-functional
File locking was non-existant in this code and fccs chained list was
not used at all. This resulted in an assertion failure when closing the
cache.  Code has been reorganized to parallel the code in krb5_fcc_resolve
for easier maintenence. 

Commented out test in t_cc.c has been updated to actually test this code.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19920 dc483132-0cff-0310-8789-dd5450dbe970