ticket: new subject: NIM: BUG: khm_krb5_initialize() failed to return error code component: windows
khm_krb5_initialize() is called in the krb5cred.dll and krb4cred.dll 
credential providers in order to ensure that the caller has references
to a valid krb5_context and a valid krb5_ccache.  If the krb5_cc_resolve()
call failed, the error code was not being returned to the caller.
Instead, success was returned which in turn would result in the caller
believing the NULL krb5_ccache pointer was in fact valid.

This fix resolves Microsoft's WER Event ID 432405961.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19967 dc483132-0cff-0310-8789-dd5450dbe970