ticket: 5766
Remove the broken optimization from UnicodeToANSI() that attempted
in a stupid way to avoid to WideCharToMultiByte() call by testing 
to see if the second byte (not wchar) was NUL.  This test works 
for Latin1 but not for anything more interesting.  Always call 
WideCharToMultiByte().  

I commented out the use of WC_NO_BEST_FIT_CHARS in order to 
provide compatibility with Windows.  Windows converts the user
name without that option.  

With the current code a principal name consisting of a single
component equivalent to the Greek character Sigma and a realm 
name will be converted to S@REALM exactly as the "WhoAmI" 
command does.  If WC_NO_BEST_FIT_CHARS was specified, this 
string would be converted to "?@REALM".



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19970 dc483132-0cff-0310-8789-dd5450dbe970