ticket: new subject: NIM: BUG: KMM: miscellaneous fixes component: windows
kmm_reg.c:
Allocate enough memory to hold the entire PluginList multi-string plus 
an extra NUL if the registry value was not properly terminated.

kmm_registrar.c:
Do not record an error loading a module if there was no configuration
for it.

kmmmain.c:
Avoid a race condition by sending the thread quit message instead of
posting it.  Do so outside of the critical section to avoid a deadlock.





git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19976 dc483132-0cff-0310-8789-dd5450dbe970