ticket: 5719
Add a "View all identities" option to the view menu of Network
Identity Manager.  This option will be toggled when selected.  If
enabled, all the known identities will be displayed in any credentials
view which uses the identity of credentials as the primary group-by
field.

The setting of this option is persistent.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19982 dc483132-0cff-0310-8789-dd5450dbe970