ticket: 5724
Add a "Set default" sub menu to the Network Identity Manager
notification icon context menu.  The submenu will display a list of
identities that the user can select as the default identity.

Each identity that is displayed in the sub menu will be colored the
same way it is colored in the basic view to provide a hint as to the
state of the credentials belonging to the identity.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19983 dc483132-0cff-0310-8789-dd5450dbe970