ticket: new subject: NIM: FEATURE: APP: Notification Icon Tooltip component: windows
The Network Identity Manager notification icon can have a tooltip
associated with it.  The tooltip will be displayed to the user if she
hovers the mouse cursor over the notification icon.  Currently, the
tooltip is used to indicate the current status of Network Identity
Manager.  This patch adds the name of the default identity to the
tooltip.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19985 dc483132-0cff-0310-8789-dd5450dbe970