ticket: 5753
When there is no ImagePath value for a module it indicates
that the module has not been installed.  Do not report an
error to the end user for uninstalled modules.  This will
only annoy the user and there is probably nothing the user
can do about it.  Perhaps the user uninstalled the module
intentionally.  The fact that there is still a HKCU key
remaining is not an indication that the module is in use.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20023 dc483132-0cff-0310-8789-dd5450dbe970