ticket: new subject: NIM: BUG: APP: leaking prompts in obtain new credentials dialog
The Kerberos v5 plug-in for Network Identity Manager was not clearing
the list of prompts properly when a user changes the active identity
in the new credentials dialog.  The stale prompts would be visible to
the user if the newly selected identity is invalid or the new identity
cannot be validated.

This patch clears the prompts if there is an identity change.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20028 dc483132-0cff-0310-8789-dd5450dbe970