ticket: 5783
The credentials display in Network Identity Manager colors the
background of identity headings based on the expiration state of the
credentials that belong to the identity.

This patch changes the behavior to color these headings based on the
existence and expiration state of identity credentials.  With this
change, only identities that have valid identity credentials will
appear in green.

The same logic is used when drawing the menu items in the "Set
default" sub-menu in the notification icon.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20032 dc483132-0cff-0310-8789-dd5450dbe970