ticket: 5782
When Network Identity Manager starts, the credentials display
enumerates known identities and displays them if necessary.  However,
as plug-ins load and the list of known identities, their properties
and credentials change, the order of the displayed list will also
change.

The existing code does not deal well in respect to tracking the
current selection when the list of identities and credentials change.
This results in inconsistent behavior where the identity that was
originally at the top of the list maintains its selection state while
the identity at the top of the list after all the plug-ins have loaded
will receive focus.

This patch fixes this problem by explicitly setting focus and
selection to the identity or credential at the current cursor location
following a change in the credentials list.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20033 dc483132-0cff-0310-8789-dd5450dbe970