ticket: new subject: remove error tables by pointer tags: pullup target_version: 1.6.3
On Windows, it is possible for the same DLL to be loaded 
into a process multiple times as separate instances.  Each
time a DLL is loaded it registers its error tables at different
locations in the process address space.  Removing the tables
by base instead of pointer value can result in the error table
list pointing at invalid memory.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20040 dc483132-0cff-0310-8789-dd5450dbe970