ticket: 5820
There appears to be a bug either in the WiX engine or the Windows Installer 3.1.
The "File" type on the Registry Search property is supposed to provide the full 
path name.  Instead, we are being given just the directory as if it were being 
processed with the "Directory" type.

We can avoid this for a REG_SZ value by using the "Raw" type because we are 
sure that the string is not going to begin with a '#' character.  

Because the full path was not being obtained for the UPGRADENSIS property, the
Uninstall routine was unable to CreateProcess() the uninstall program.  

This commit also includes addition debugging in the NSIS Uninstall custom
handler to report the path and the GetLastError() value when the uninstall 
fails.  This will be logged in the msiexec log file and displayed in a 
MessageBox.




git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20119 dc483132-0cff-0310-8789-dd5450dbe970