ticket: new subject: Windows: avoid use of cygwin mkdir and rmdir commands tags: pullup
Microsoft's nmake versions 8.x and 9.x prefer executables over
internal shell commands.  This is a change from previous versions.
Cygwin's mkdir and rmdir commands do not have the same semantics
as the cmd.exe shell versions.  

Change the definitions of MKDIR and RMDIR to use 'md' and 'rd'
in order to avoid the use of the cygwin versions.




git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20211 dc483132-0cff-0310-8789-dd5450dbe970