ticket: new status: open subject: cc_ccache_iterator_release, cc_credentials_iterator leak server memory
cc_ccache_iterator_release and cc_credentials_iterator leak memory on the 
CCacheServer because they do not send an ipc message to the server telling it
to free up resources associated with the iterator.

Note: this issue does not track needing to be able to free up these same 
resources when the client crashes. 


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20276 dc483132-0cff-0310-8789-dd5450dbe970